Renaming a few things to be less wordy
